Support Delegated Proving Across All Transaction Types
