+++ crates/rust-client/src/note/import.rs
@@ -259,7 +259,7 @@ impl<R: FeltRng> Client<R> {
         loop {
             if request_block_num > current_block_num {
                 return Ok(None);
-            };
+            }
 
             let sync_notes = self.rpc_api.sync_notes(request_block_num, &[tag]).await?;
 
@@ -279,7 +279,7 @@ impl<R: FeltRng> Client<R> {
 
                 if note_block_num > current_block_num {
                     return Ok(None);
-                };
+                }
 
                 let note_inclusion_proof = NoteInclusionProof::new(
                     note_block_num,
