+++ crates/web-client/js/constants.js
@@ -8,9 +8,6 @@ export const MethodName = Object.freeze({
   NEW_WALLET: "newWallet",
   NEW_FAUCET: "newFaucet",
   NEW_TRANSACTION: "newTransaction",
-  NEW_MINT_TRANSACTION: "newMintTransaction",
-  NEW_CONSUME_TRANSACTION: "newConsumeTransaction",
-  NEW_SEND_TRANSACTION: "newSendTransaction",
   SUBMIT_TRANSACTION: "submitTransaction",
   SYNC_STATE: "syncState",
 });
