+++ crates/web-client/js/index.js
@@ -326,104 +326,6 @@ export class WebClient {
     }
   }
 
-  async newMintTransaction(targetAccountId, faucetId, noteType, amount) {
-    try {
-      if (!this.worker) {
-        return await this.wasmWebClient.newMintTransaction(
-          targetAccountId,
-          faucetId,
-          noteType,
-          amount
-        );
-      }
-      const serializedTargetAccountId = targetAccountId.toString();
-      const serializedFaucetId = faucetId.toString();
-      const serializedNoteType = noteType.serialize();
-      const serializedAmount = amount.toString();
-      const serializedTransactionResultBytes = await this.callMethodWithWorker(
-        MethodName.NEW_MINT_TRANSACTION,
-        serializedTargetAccountId,
-        serializedFaucetId,
-        serializedNoteType,
-        serializedAmount
-      );
-      return wasm.TransactionResult.deserialize(
-        new Uint8Array(serializedTransactionResultBytes)
-      );
-    } catch (error) {
-      console.error("INDEX.JS: Error in newMintTransaction:", error);
-      throw error; // Ensure the test catches and asserts
-    }
-  }
-
-  async newConsumeTransaction(targetAccountId, noteId) {
-    try {
-      if (!this.worker) {
-        return await this.wasmWebClient.newConsumeTransaction(
-          targetAccountId,
-          noteId
-        );
-      }
-      const serializedTargetAccountId = targetAccountId.toString();
-      const serializedTransactionResultBytes = await this.callMethodWithWorker(
-        MethodName.NEW_CONSUME_TRANSACTION,
-        serializedTargetAccountId,
-        noteId
-      );
-      return wasm.TransactionResult.deserialize(
-        new Uint8Array(serializedTransactionResultBytes)
-      );
-    } catch (error) {
-      console.error(
-        "INDEX.JS: Error in newConsumeTransaction:",
-        JSON.stringify(error)
-      );
-      throw error;
-    }
-  }
-
-  async newSendTransaction(
-    senderAccountId,
-    receiverAccountId,
-    faucetId,
-    noteType,
-    amount,
-    recallHeight = null
-  ) {
-    try {
-      if (!this.worker) {
-        return await this.wasmWebClient.newSendTransaction(
-          senderAccountId,
-          receiverAccountId,
-          faucetId,
-          noteType,
-          amount,
-          recallHeight
-        );
-      }
-      const serializedSenderAccountId = senderAccountId.toString();
-      const serializedReceiverAccountId = receiverAccountId.toString();
-      const serializedFaucetId = faucetId.toString();
-      const serializedNoteType = noteType.serialize();
-      const serializedAmount = amount.toString();
-      const serializedTransactionResultBytes = await this.callMethodWithWorker(
-        MethodName.NEW_SEND_TRANSACTION,
-        serializedSenderAccountId,
-        serializedReceiverAccountId,
-        serializedFaucetId,
-        serializedNoteType,
-        serializedAmount,
-        recallHeight
-      );
-      return wasm.TransactionResult.deserialize(
-        new Uint8Array(serializedTransactionResultBytes)
-      );
-    } catch (error) {
-      console.error("INDEX.JS: Error in newSendTransaction:", error);
-      throw error;
-    }
-  }
-
   async submitTransaction(transactionResult, prover = undefined) {
     try {
       if (!this.worker) {
